requirements.txt
fastapi[all]
uvicorn
sqlalchemy
asyncpg
pydantic
alembic
python-dotenv
bcrypt
passlib[bcrypt]
pyjwt
